Download maps and plans from the 50-State Simulation Project

These functions will download redist_map and redist_plans objects for the 50-State Simulation Project from the ALARM Project's Dataverse. alarm_50state_doc() will download documentation for a particular state and show it in a browser. alarm_50state_stats will download just the summary statistics for a state.

Usage

alarm_50state_map(state, year = 2020)

alarm_50state_plans(state, stats = TRUE, year = 2020)

alarm_50state_stats(state, year = 2020)

alarm_50state_doc(state, year = 2020)

Arguments

state

A state name, abbreviation, FIPS code, or ANSI code.

year

The redistricting cycle to download. Currently only "2020" is available.

stats

if TRUE (the default), download summary statistics for each plan.

Value

For alarm_50state_map(), a redist_map. For alarm_50state_plans(), a redist_plans. For alarm_50state_doc(), nothing (but load an HTML file into the viewer or web browser). For alarm_50state_stats(), a tibble.

Details

Every decade following the Census, states and municipalities must redraw districts for Congress, state houses, city councils, and more. The goal of the 50-State Simulation Project is to enable researchers, practitioners, and the general public to use cutting-edge redistricting simulation analysis to evaluate enacted congressional districts.

Evaluating a redistricting plan requires analysts to take into account each state’s redistricting rules and particular political geography. Comparing the partisan bias of a plan for Texas with the bias of a plan for New York, for example, is likely misleading. Comparing a state’s current plan to a past plan is also problematic because of demographic and political changes over time. Redistricting simulations generate an ensemble of alternative redistricting plans within a given state which are tailored to its redistricting rules. Unlike traditional evaluation methods, therefore, simulations are able to directly account for the state’s political geography and redistricting criteria.

Examples

if (FALSE) {
# requires stable connection to the Harvard Dataverse
alarm_50state_map("WA")
alarm_50state_plans("WA", stats=FALSE)
alarm_50states_stats('WA')
alarm_50state_doc("WA")
}

alarm_50state_plans('WY')
#> An earlier bug incorrectly in redist doubled the number of removed edges.
#> Current counts are correct.
#> This message is displayed once per session.
#> redist 4.0.0 aligned pro-Republican biases as positive numbers, reversing the
#> old Efficiency Gap direction.
#> This message is displayed once per session.
#> A <redist_plans> containing 5000 sampled plans and 1 reference plan
#> Plans have 1 districts from a 1-unit map, and were drawn using an unknown
#> algorithm.
#> Plans matrix: int [1, 1:5001] 1 1 1 1 1 1 1 1 1 1 ...
#> # A tibble: 5,001 × 49
#>    draw  distr…¹ total…² total…³ plan_…⁴ comp_…⁵ comp_…⁶ pop_h…⁷ pop_w…⁸ pop_b…⁹
#>    <fct>   <int>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 cd_2…       1  576851  441337       0     NaN   0.772   59046  469664    4735
#>  2 1           1  576851  441337       0     NaN   0.772   59046  469664    4735
#>  3 2           1  576851  441337       0     NaN   0.772   59046  469664    4735
#>  4 3           1  576851  441337       0     NaN   0.772   59046  469664    4735
#>  5 4           1  576851  441337       0     NaN   0.772   59046  469664    4735
#>  6 5           1  576851  441337       0     NaN   0.772   59046  469664    4735
#>  7 6           1  576851  441337       0     NaN   0.772   59046  469664    4735
#>  8 7           1  576851  441337       0     NaN   0.772   59046  469664    4735
#>  9 8           1  576851  441337       0     NaN   0.772   59046  469664    4735
#> 10 9           1  576851  441337       0     NaN   0.772   59046  469664    4735
#> # … with 4,991 more rows, 39 more variables: pop_aian <dbl>, pop_asian <dbl>,
#> #   pop_nhpi <dbl>, pop_other <dbl>, pop_two <dbl>, vap_hisp <dbl>,
#> #   vap_white <dbl>, vap_black <dbl>, vap_aian <dbl>, vap_asian <dbl>,
#> #   vap_nhpi <dbl>, vap_other <dbl>, vap_two <dbl>, pre_16_rep_tru <dbl>,
#> #   pre_16_dem_cli <dbl>, uss_18_rep_bar <dbl>, uss_18_dem_tra <dbl>,
#> #   gov_18_rep_gor <dbl>, gov_18_dem_thr <dbl>, sos_18_rep_buc <dbl>,
#> #   sos_18_dem_byr <dbl>, pre_20_rep_tru <dbl>, pre_20_dem_bid <dbl>, …
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
alarm_50state_stats("WY")
#> # A tibble: 5,001 × 49
#>    draw  distr…¹ total…² total…³ plan_…⁴ comp_…⁵ comp_…⁶ pop_h…⁷ pop_w…⁸ pop_b…⁹
#>    <fct>   <int>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 cd_2…       1  576851  441337       0     NaN   0.772   59046  469664    4735
#>  2 1           1  576851  441337       0     NaN   0.772   59046  469664    4735
#>  3 2           1  576851  441337       0     NaN   0.772   59046  469664    4735
#>  4 3           1  576851  441337       0     NaN   0.772   59046  469664    4735
#>  5 4           1  576851  441337       0     NaN   0.772   59046  469664    4735
#>  6 5           1  576851  441337       0     NaN   0.772   59046  469664    4735
#>  7 6           1  576851  441337       0     NaN   0.772   59046  469664    4735
#>  8 7           1  576851  441337       0     NaN   0.772   59046  469664    4735
#>  9 8           1  576851  441337       0     NaN   0.772   59046  469664    4735
#> 10 9           1  576851  441337       0     NaN   0.772   59046  469664    4735
#> # … with 4,991 more rows, 39 more variables: pop_aian <dbl>, pop_asian <dbl>,
#> #   pop_nhpi <dbl>, pop_other <dbl>, pop_two <dbl>, vap_hisp <dbl>,
#> #   vap_white <dbl>, vap_black <dbl>, vap_aian <dbl>, vap_asian <dbl>,
#> #   vap_nhpi <dbl>, vap_other <dbl>, vap_two <dbl>, pre_16_rep_tru <dbl>,
#> #   pre_16_dem_cli <dbl>, uss_18_rep_bar <dbl>, uss_18_dem_tra <dbl>,
#> #   gov_18_rep_gor <dbl>, gov_18_dem_thr <dbl>, sos_18_rep_buc <dbl>,
#> #   sos_18_dem_byr <dbl>, pre_20_rep_tru <dbl>, pre_20_dem_bid <dbl>, …
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names