alarmdata: Download, Merge, and Process Redistricting Data

Utility functions to download and process data produced by the ALARM Project, including 2020 redistricting files Kenny and McCartan (2021) https://alarm-redist.org/posts/2021-08-10-census-2020/ and the 50-State Redistricting Simulations of McCartan, Kenny, Simko, Garcia, Wang, Wu, Kuriwaki, and Imai (2022) doi:10.7910/DVN/SLCD3E . The package extends the data introduced in McCartan, Kenny, Simko, Garcia, Wang, Wu, Kuriwaki, and Imai (2022) doi:10.1038/s41597-022-01808-2 to also include states with only a single district. The package also includes the Japanese 2022 redistricting files from the 47-Prefecture Redistricting Simulations of Miyazaki, Yamada, Yatsuhashi, and Imai (2022) doi:10.7910/DVN/Z9UKSH .

See also

Author

Maintainer: Cory McCartan mccartan@psu.edu (ORCID)

Authors:

Christopher T. Kenny christopherkenny@fas.harvard.edu (ORCID)

Tyler Simko tsimko@g.harvard.edu (ORCID)

Michael Zhao michaelzhao@college.harvard.edu (ORCID)

Sho Miyazaki shomiya@stanford.edu (ORCID)

Kosuke Imai imai@harvard.edu (ORCID)