Add a reference plan to a set of plans

Facilitates comparing an existing (i.e., non-simulated) redistricting plan to a set of simulated plans.

Usage

alarm_add_plan(
  plans,
  ref_plan,
  map = NULL,
  name = NULL,
  calc_polsby = FALSE,
  GEOID = "GEOID"
)

Arguments

plans

A redist_plans object.

ref_plan

An integer vector containing the reference plan or a block assignment file as a tibble or data.frame.

map

A redist_map object. Only required if the redist_plans object includes summary statistics.

name

A human-readable name for the reference plan. Defaults to the name of ref_plan. If ref_plan is a tibble or data.frame, it should be the name of the column of ref_plan that identifies districts.

calc_polsby

A logical value indicating whether a Polsby-Popper compactness score should be calculated for the reference plan. Defaults to FALSE.

GEOID

character. Ignored unless ref_plan is a tibble or data.frame. Should correspond to the column of ref_plan that identifies block GEOIDs. Default is 'GEOID'.

Value

A modified redist_plans object containing the reference plan. Includes summary statistics if the original redist_plans object had them as well.

Examples

map <- alarm_50state_map("WY")
pl <- alarm_50state_plans("WY")
alarm_add_plan(pl, ref_plan = c(1), map, name = "example")
#> Warning: There were 2 warnings in `"draw" %in% names(data)`.
#> The first warning was:
#> ℹ In argument: `comp_edge = redist::distr_compactness(map)`.
#> Caused by warning:
#> ! 'redist::distr_compactness' is deprecated.
#> See help("Deprecated")
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 4 warnings in `"draw" %in% names(data)`.
#> The first warning was:
#> ℹ In argument: `egap = redist::partisan_metrics(map, "EffGap", rvote, dvote)`.
#> Caused by warning:
#> ! 'redist::partisan_metrics' is deprecated.
#> See help("Deprecated")
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 3 remaining warnings.
#> Warning: There were 4 warnings in `"draw" %in% names(data)`.
#> The first warning was:
#> ℹ In argument: `egap = redist::partisan_metrics(map, "EffGap", rvote, dvote)`.
#> Caused by warning:
#> ! 'redist::partisan_metrics' is deprecated.
#> See help("Deprecated")
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 3 remaining warnings.
#> Warning: There were 4 warnings in `"draw" %in% names(data)`.
#> The first warning was:
#> ℹ In argument: `egap = redist::partisan_metrics(map, "EffGap", rvote, dvote)`.
#> Caused by warning:
#> ! 'redist::partisan_metrics' is deprecated.
#> See help("Deprecated")
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 3 remaining warnings.
#> Warning: There were 4 warnings in `"draw" %in% names(data)`.
#> The first warning was:
#> ℹ In argument: `egap = redist::partisan_metrics(map, "EffGap", rvote, dvote)`.
#> Caused by warning:
#> ! 'redist::partisan_metrics' is deprecated.
#> See help("Deprecated")
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 3 remaining warnings.
#> Warning: There were 4 warnings in `"draw" %in% names(data)`.
#> The first warning was:
#> ℹ In argument: `egap = redist::partisan_metrics(map, "EffGap", rvote, dvote)`.
#> Caused by warning:
#> ! 'redist::partisan_metrics' is deprecated.
#> See help("Deprecated")
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 3 remaining warnings.
#> Warning: There were 4 warnings in `"draw" %in% names(data)`.
#> The first warning was:
#> ℹ In argument: `egap = redist::partisan_metrics(map, "EffGap", rvote, dvote)`.
#> Caused by warning:
#> ! 'redist::partisan_metrics' is deprecated.
#> See help("Deprecated")
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 3 remaining warnings.
#> A <redist_plans> containing 5,000 sampled plans and 2 reference plans
#> Plans have 1 district from a 1-unit map, and were drawn using an unknown
#> algorithm.
#> Plans matrix: int [1, 1:5002] 1 1 1 1 1 1 1 1 1 1 ...
#> # A tibble: 5,002 × 50
#>    chain draw    district total_pop total_vap plan_dev comp_edge comp_polsby
#>    <int> <fct>      <int>     <dbl>     <dbl>    <dbl>     <dbl>       <dbl>
#>  1     1 example        1    576851    441337        0       NaN      NA    
#>  2     2 cd_2020        1    576851    441337        0       NaN       0.772
#>  3     2 1              1    576851    441337        0       NaN       0.772
#>  4     2 2              1    576851    441337        0       NaN       0.772
#>  5     2 3              1    576851    441337        0       NaN       0.772
#>  6     2 4              1    576851    441337        0       NaN       0.772
#>  7     2 5              1    576851    441337        0       NaN       0.772
#>  8     2 6              1    576851    441337        0       NaN       0.772
#>  9     2 7              1    576851    441337        0       NaN       0.772
#> 10     2 8              1    576851    441337        0       NaN       0.772
#> # ℹ 4,992 more rows
#> # ℹ 42 more variables: pop_hisp <dbl>, pop_white <dbl>, pop_black <dbl>,
#> #   pop_aian <dbl>, pop_asian <dbl>, pop_nhpi <dbl>, pop_other <dbl>,
#> #   pop_two <dbl>, vap_hisp <dbl>, vap_white <dbl>, vap_black <dbl>,
#> #   vap_aian <dbl>, vap_asian <dbl>, vap_nhpi <dbl>, vap_other <dbl>,
#> #   vap_two <dbl>, pre_16_rep_tru <dbl>, pre_16_dem_cli <dbl>,
#> #   uss_18_rep_bar <dbl>, uss_18_dem_tra <dbl>, gov_18_rep_gor <dbl>, …

if (FALSE) {
# requires Harvard Dataverse API key
# download and load a comparison plan
url <- "https://www.redistrict2020.org/files/NM-2021-10/Congressional_Concept_A.zip"
tf <- tempfile(fileext = ".zip")
utils::download.file(url, tf)
utils::unzip(tf, exdir = dirname(tf))
baf <- readr::read_csv(file = paste0(dirname(tf), "/Congressional Concept A.csv"),
                       col_types = "ci")
names(baf) <- c("GEOID", "concept_a")
# Add it to the plans object
map_nm <- alarm_50state_map("NM")
plans_nm <- alarm_50state_plans("NM", stats = FALSE)
alarm_add_plan(plans_nm, baf, map = map_nm, name = "concept_a")
}